% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eva_cal.R
\name{eva_cal}
\alias{eva_cal}
\title{eva_cal}
\usage{
eva_cal(prediction, cell_label)
}
\arguments{
\item{prediction}{A vector of annotate cell type labels}

\item{cell_label}{A vector of original cell type labels}
}
\value{
A matrix contain the F1 score of each cell population, mean of F1 score and overall accuracy
}
\description{
calculate the F1 score of each cell population, mean of F1 score and overall accuracy
}
\examples{
data(predict_label)
data(pbmc2)
eva_cal(prediction = predict_label,cell_label = pbmc2[,1])

}
