% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc2}
\alias{pbmc2}
\title{pbmc2}
\format{
a data frame
}
\usage{
data(pbmc2, package="scAnnotate")
}
\description{
A subset of human PBMC scRNA-seq data that was sequenced using inDrops platform.
The Seurat(version 4.0.5) package was used for normalized using the NormalizeData
function with the "LogNormalize" method and a scale factor of 10,000. After modeling
the mean-variance relationship with the  FindVariableFeautre function within "vst"
methods, we selected the top 2,000 highly variable genes and only used this selection
going forward. The dataframe of the cell type label and a gene expression matrix of
644 cells in the row and 2,000 genes in the column.
}
\references{
Ding,  J.et  al.(2019).   Systematic  comparative  analysis  of  single  cellrna-sequencing methods.bioRxiv
}
\keyword{datasets}
