% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{predict_label}
\alias{predict_label}
\title{predict_label}
\format{
a data frame
}
\usage{
data(predict_label, package="scAnnotate")
}
\description{
Cell type annotation of pbmc2 data that training from pbmc1 data by 'scAnnotate'.
}
\keyword{datasets}
